+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Collection View is now compatible with Thingworx 8.4 and compiled in strict mode.
 
+A new special `@row` property can now be used as a binding source and represents the entire infotable row.
+
 Table Layout has been deprecated and cannot be selected as a layout for newly created collection views. Old collection views using it will continue to function, but the option is no longer visible in the layout options - **changing the layout for one of these collection views will be irreversible**. Flow Layout with the `FlowLayoutMaximumCellsPerRow` property set to `1` can now be used instead of table layout to obtain the same behaviour with more options for configuration.
 
 Collection View widget will now behave as a view-based widget when part of a view hierarchy - in other cases it will behave as a regular widget. When part of a view hierarchy, the `coreUIView` property will return the collection view.